APIAuthorizationListenerService deleted and distributed function among classes
