+++ src/Serializer/ItemNormalizer.php
@@ -16,13 +16,13 @@ final class ItemNormalizer
     public function normalize(
         Item $item,
         string $format = null,
-        array $context = []
+        string $context = null
     ): array|null
     {
         return [
             'id' => $item->getId(),
             'name' => $item->getName(),
-            'project' => is_null($item->getProject()) ? null : $this->projectNormalizer->normalize($item->getProject(), null, in_array('pagination', $context) ? ['pagination'] : []),
+            'project' => is_null($item->getProject()) ? null : $this->projectNormalizer->normalize($item->getProject(), null, $context),
             'parameter' => json_decode($item->getParameter(), true),
             'path' => json_decode($item->getPath(), true),
             'creationDate' => $item->getCreationDate(),
@@ -36,7 +36,7 @@ final class ItemNormalizer
     public function supportsNormalization(
         $data,
         string $format = null,
-        array $context = []
+        string $context = null
     ): bool
     {
         return $data instanceof Item;
