+++ src/Service/Listener/APIInventoryListenerService.php
@@ -1,0 +1,147 @@
+<?php declare(strict_types=1);
+
+namespace App\Service\Listener;
+
+use App\Entity\User;
+use App\Service\API\UserService;
+use App\Service\API\Item\ItemService;
+use App\Repository\InventoryRepository;
+use Symfony\Contracts\Cache\ItemInterface;
+use Symfony\Contracts\Cache\CacheInterface;
+use App\Service\Response\API\CustomResponse;
+use App\Service\API\Security\SecurityService;
+use Symfony\Component\HttpKernel\Event\RequestEvent;
+
+final class APIInventoryListenerService
+{
+    public function __construct(
+        private readonly InventoryRepository $inventoryRepository,
+        private readonly SecurityService $securityService,
+        private readonly CustomResponse $customResponse,
+        private readonly ItemService $itemService,
+        private readonly UserService $userService,
+        private readonly CacheInterface $cache
+    )
+    {
+    }
+
+    public function validateJWTForInventoryController(
+        RequestEvent $event,
+        array $accessToken,
+        array $params
+    ): void
+    {
+        if ($this->isRouteInventories($event, $accessToken, $params)) {
+            return;
+        }
+
+        $uuid = $params['uuid'];
+
+        $user = $this->userService->getUserByUuidFromCache($uuid);
+
+        if (is_null($user)) {
+            $event->setResponse($this->customResponse->errorResponse($event->getRequest(), sprintf('User with uuid %s doesn\'t exists!', $uuid), 404));
+
+            return;
+        }
+
+        if ($this->isRouteInventoryByUuid($event, $accessToken, $params, $user)) {
+            return;
+        }
+
+        $itemId = $params['itemId'];
+
+        $item = json_decode($this->itemService->getItemFromCacheById($itemId), true);
+
+        if (($user->isPrivate() && !$this->securityService->hasClientPermissionForAdjustmentOnUserInventory($accessToken, $user, $item) || !$user->isPrivate() && !$event->getRequest()->isMethod('GET') && !$this->securityService->hasClientPermissionForAdjustmentOnUserInventory($accessToken, $user, $item))) {
+            $event->setResponse($this->customResponse->errorResponse($event->getRequest(), 'Permission denied!', 403));
+
+            return;
+        }
+
+        if (is_null($item)) {
+            $event->setResponse($this->customResponse->errorResponse($event->getRequest(), 'Item not found', 404));
+
+            return;
+        }
+
+        if ($event->getRequest()->isMethod('GET')) {
+            $this->cache->get('inventory_' . $uuid . '_item_' . $itemId, function (ItemInterface $cacheItem) use ($user, $item) {
+                $cacheItem->expiresAfter(86400);
+
+                return $this->inventoryRepository->findOneBy(['user' => $user->getId(), 'item' => $item]);
+            });
+
+            return;
+        }
+
+        $inventory = $this->inventoryRepository->findOneBy(['user' => $user->getId(), 'item' => $item]);
+
+        if (is_null($inventory) && $event->getRequest()->isMethod('PATCH')) {
+            $event->setResponse($this->customResponse->errorResponse($event->getRequest(), sprintf('User has no item with id %s in inventory yet! Please use POST method to add item!', $itemId), 406));
+
+            return;
+        }
+
+        if (!is_null($inventory) && $event->getRequest()->isMethod('POST')) {
+            $event->setResponse($this->customResponse->errorResponse($event->getRequest(), sprintf('User already has item with id %s in inventory. For update use PATCH method', $itemId), 406));
+
+            return;
+        }
+
+        if (is_null($inventory) && !$event->getRequest()->isMethod('POST')) {
+            $event->setResponse($this->customResponse->errorResponse($event->getRequest(), sprintf('User has no item with id %s in inventory yet!', $itemId), 406));
+        }
+    }
+
+    private function isRouteInventories(
+        RequestEvent $event,
+        array $accessToken,
+        array $params
+    ): bool
+    {
+        if (count($params) !== 0) {
+            return false;
+        }
+
+        if (!$this->securityService->isClientAdmin($accessToken)) {
+            $event->setResponse($this->customResponse->errorResponse($event->getRequest(), 'Permission denied!', 403));
+
+            return true;
+        }
+
+        return true;
+    }
+
+    private function isRouteInventoryByUuid(
+        RequestEvent $event,
+        array $accessToken,
+        array $params,
+        User $user
+    ): bool
+    {
+        if (count($params) > 1) {
+            return false;
+        }
+
+        if ($user->isPrivate() && !$this->securityService->hasClientPermissionForAccessingUserRelatedData($accessToken, $user)) {
+            $event->setResponse($this->customResponse->errorResponse($event->getRequest(), 'Permission denied!', 403));
+
+            return true;
+        }
+
+        $inventory = $this->cache->get('inventory_' . $user->getUuid(), function (ItemInterface $item) use ($user) {
+            $item->expiresAfter(86400);
+
+            return $this->inventoryRepository->findBy(['user' => $user->getUuid()]);
+        });
+
+        if (count($inventory) === 0) {
+            $event->setResponse($this->customResponse->errorResponse($event->getRequest(), 'User has not an item in inventory yet!', 404));
+        }
+
+        return true;
+    }
+
+
+}
